Simplify this whole tool
There is NO NEED to build MavenProject, that in fact is not
even part of Maven public API. Moreover, the reason why
MavenProject is built is really just to grab "raw" and "effective"
models, that is in fact model builder job to do.

Clean up several other issues as well, drop the huge amount
of not needed and/or unused dependencies as well.